+++ src/DataStructures/InfluxDatabase.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -18,12 +19,33 @@ namespace AdysTech.InfluxDB.Client.Net
             }
         }
 
+        public ICollection<IInfluxRetentionPolicy> RetentionPolicies
+        {
+            get
+            {
+                return new ReadOnlyCollection<IInfluxRetentionPolicy>(_rps.Keys.ToList());
+            }
+        }
+
+        public IDictionary<IInfluxRetentionPolicy, ICollection<IInfluxMeasurement>> MeasurementHierarchy
+        {
+            get
+            {
+                return _rps;
+            }
+        }
+
         HashSet<IInfluxMeasurement> _measurements;
+        Dictionary<IInfluxRetentionPolicy, ICollection<IInfluxMeasurement>> _rps;
+
         public InfluxDatabase(string name)
         {
             Name = name;
-            _measurements  = new HashSet<IInfluxMeasurement>();
+            _measurements = new HashSet<IInfluxMeasurement>();
+            _rps = new Dictionary<IInfluxRetentionPolicy, ICollection<IInfluxMeasurement>>();
         }
+
+
 
     }
 }