+++ src/Interfaces/IInfluxDatabase.cs
@@ -4,6 +4,8 @@ namespace AdysTech.InfluxDB.Client.Net
 {
     public interface IInfluxDatabase
     {
+        IDictionary<IInfluxRetentionPolicy, ICollection<IInfluxMeasurement>> MeasurementHierarchy{ get; }
+
         ICollection<IInfluxMeasurement> Measurements { get; }
         string Name { get; }
     }