+++ src/Interfaces/IInfluxMeasurement.cs
@@ -7,5 +7,8 @@ namespace AdysTech.InfluxDB.Client.Net
         string Name { get; }
         ICollection<string> Fields { get; }
         ICollection<string> Tags { get; }
+
+        int SeriesCount { get; }
+        int PointsCount { get; }
     }
 }